app/backend/getaway/requirements.txt
uvicorn
pydantic
pydantic_settings
fastapi
sqlalchemy
grpcio
grpcio-tools
protobuf
redis
httpx